fqdn: move defaultdns from pkg/proxy to pkg/fqdn
PR https://github.com/cilium/cilium/pull/38020 extracted the default dns
singleton into a separate cell in package `pkg/proxy/defaultdns`.

But package `pkg/fqdn` would actually be the better parent package
as `pkg/proxy` is only a proxy layer that shouldn't contain actual
L7 proxy logic (Envoy & FQDN).

Therefore, this commit moves the new cell `defaultdns` from `pkg/proxy/defaultdns`
into `pkg/fqdn/defaultdns`.
